Cast: Cleaning up static_casts
Changed to use size_t through out the code to avoid 
static_casts. Also added a few DCHECKs in the locations 
where we really downcast to a smaller size.

Review URL: https://codereview.chromium.org/26253007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@229909 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/rtp_receiver/rtp_parser/rtp_parser_unittest.cc
@@ -14,8 +14,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace media {
 namespace cast {
 
-static const int kPacketLength = 1500;
-static const int kCastRtpLength = 7;
+static const size_t kPacketLength = 1500;
+static const size_t kCastRtpLength = 7;
 static const int kTestPayloadType = 127;
 static const uint32 kTestSsrc = 1234;
 static const uint32 kTestTimestamp = 111111;
@@ -35,7 +35,7 @@ class RtpDataTest : public RtpData {
   }
 
   virtual void OnReceivedPayloadData(const uint8* payloadData,
-                                     int payloadSize,
+                                     size_t payloadSize,
                                      const RtpCastHeader* rtpHeader) OVERRIDE {
     VerifyCommonHeader(*rtpHeader);
     VerifyCastHeader(*rtpHeader);
